requirements.txt
requests
langchain
openai
regex
pyyaml
instructor
llama-index
flask